Use spawnOn to map startup apps to a workspace
Previously workspace manage hooks were being used to assign particular
applications to a particular workspace when starting up.  This was lame
because sometimes an application might be on multiple workspaces (e.g.
a terminal).  This commit switches things to use 'spawnOn' to assign
applications on startup instead.

Also switched the default browser to Chromium.